requirements.txt
Django
djangorestframework
django-rest-auth
psycopg2
django-allauth
flake8
argon2-cffi
django[argon2]
